Test about execute two kernel in sequence.

+++ shader.cl
@@ -1,4 +1,4 @@
-__kernel void adder(__global const float* a, __global const float* b, __global float* result)
+__kernel void adder(__global const int* a, __global const int* b, __global int* result)
 {
 	int idx = get_global_id(0);
 	result[idx] = a[idx] + b[idx];
